+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [6.49.0-rc.4](https://github.com/consolelabs/mochi-discord/compare/v6.49.0-rc.3...v6.49.0-rc.4) (2024-05-21)
+
+
+### Bug Fixes
+
+* update usd format for watch list view ([#1662](https://github.com/consolelabs/mochi-discord/issues/1662)) ([be3db6f](https://github.com/consolelabs/mochi-discord/commit/be3db6fb20fef75565621881939631a4d9bfd02a))
+
+# [6.49.0-rc.3](https://github.com/consolelabs/mochi-discord/compare/v6.49.0-rc.2...v6.49.0-rc.3) (2024-05-21)
+
+
+### Bug Fixes
+
+* add momentjs ([#1661](https://github.com/consolelabs/mochi-discord/issues/1661)) ([db13fa4](https://github.com/consolelabs/mochi-discord/commit/db13fa4ef30b708c9e3a8a370251df2f73343d74))
+
+# [6.49.0-rc.2](https://github.com/consolelabs/mochi-discord/compare/v6.49.0-rc.1...v6.49.0-rc.2) (2024-05-20)
+
+
+### Bug Fixes
+
+* vault detail render ([#1660](https://github.com/consolelabs/mochi-discord/issues/1660)) ([94c6443](https://github.com/consolelabs/mochi-discord/commit/94c64434c82b3e07d30fd18c862ecad92c6c3359))
+
+# [6.49.0-rc.1](https://github.com/consolelabs/mochi-discord/compare/v6.48.1...v6.49.0-rc.1) (2024-05-20)
+
+
+### Features
+
+* trading vault ([#1658](https://github.com/consolelabs/mochi-discord/issues/1658)) ([d9b1764](https://github.com/consolelabs/mochi-discord/commit/d9b1764c24d872295121fe906dc7add45e5ca008))
+
 ## [6.48.1](https://github.com/consolelabs/mochi-discord/compare/v6.48.0...v6.48.1) (2024-05-13)